feat: add uuid for table & context scheme change preview endpoint

+++ lib/Service/TableService.php
@@ -16,6 +16,7 @@ use OCA\Tables\AppInfo\Application;
 use OCA\Tables\Db\Table;
 use OCA\Tables\Db\TableMapper;
 use OCA\Tables\Dto\Column as ColumnDto;
+use OCA\Tables\Errors\BadRequestError;
 use OCA\Tables\Errors\InternalError;
 use OCA\Tables\Errors\NotFoundError;
 use OCA\Tables\Errors\PermissionError;
@@ -26,8 +27,10 @@ use OCA\Tables\Model\ColumnSettings;
 use OCA\Tables\Model\Permissions;
 use OCA\Tables\Model\SortRuleSet;
 use OCA\Tables\Model\TableScheme;
+use OCA\Tables\Model\ViewUpdateInput;
 use OCA\Tables\ResponseDefinitions;
 use OCA\Tables\Service\ValueObject\Title;
+use OCA\Tables\Vendor\Symfony\Component\Uid\Uuid;
 use OCP\App\IAppManager;
 use OCP\AppFramework\Db\DoesNotExistException;
 use OCP\AppFramework\Db\MultipleObjectsReturnedException;
@@ -37,6 +40,8 @@ use OCP\Defaults;
 use OCP\EventDispatcher\IEventDispatcher;
 use OCP\IDBConnection;
 use OCP\IL10N;
+use Psr\Container\ContainerExceptionInterface;
+use Psr\Container\NotFoundExceptionInterface;
 use Psr\Log\LoggerInterface;
 
 /**
@@ -281,6 +286,7 @@ class TableService extends SuperService {
 
 		$time = new DateTime();
 		$item = new Table();
+		$item->setUuid(null);
 		$item->setTitle($title);
 		$item->setDescription($description);
 		if ($emoji) {
@@ -571,7 +577,7 @@ class TableService extends SuperService {
 		$table = $this->find($id, skipTableEnhancement: true);
 		$columns = $this->columnService->findAllByTable($id, null, $table);
 		$this->enhanceTable($table, $userId ?? $this->userId);
-		return new TableScheme($table->getTitle(), $table->getEmoji(), $columns, $table->getViews() ?: [], $table->getDescription() ?: '', $this->appManager->getAppVersion('tables'), $table->getColumnOrderSettingsArray(), $table->getSortArray());
+		return new TableScheme($table->getTitle(), $table->getEmoji(), $columns, $table->getViews() ?: [], $table->getDescription() ?: '', $this->appManager->getAppVersion('tables'), $table->getColumnOrderSettingsArray(), $table->getSortArray(), $table->getUuid());
 	}
 
 	// PRIVATE FUNCTIONS ---------------------------------------------------------------
@@ -595,6 +601,7 @@ class TableService extends SuperService {
 	 */
 	public function importTable(array $table, string $userId): Table {
 		$item = new Table();
+		$item->setUuid((isset($table['uuid']) && Uuid::isValid($table['uuid'])) ? $table['uuid'] : null);
 		$item->setTitle($table['title']);
 		$item->setEmoji($table['emoji']);
 		$item->setOwnership($userId);
@@ -614,20 +621,66 @@ class TableService extends SuperService {
 	}
 
 	/**
-	 * @param int $tableId
-	 * @param array $addColumns
-	 * @param array $removeColumns
-	 * @param array $updateColumns
-	 * @param string|null $userId
+	 * @param int $id
+	 * @param array $updateScheme
 	 *
-	 * @return void
+	 * @return array
 	 *
 	 * @throws InternalError
 	 * @throws NotFoundError
 	 * @throws PermissionError
-	 * @throws \OCA\Tables\Errors\BadRequestError
+	 * @throws ContainerExceptionInterface
+	 * @throws NotFoundExceptionInterface
 	 */
-	public function updateTableStructure(int $tableId, array $addColumns, array $removeColumns, array $updateColumns, ?string $userId = null): void
+	public function compareTableSchemeChanges(int $id, array $updateScheme): array {
+		$table = $this->find($id);
+		$structureService = \OCP\Server::get(StructureService::class);
+		$structureService->resolveChangesForTable($id, $updateScheme);
+
+		return [
+			'title' => [
+				'from' => $table->getTitle(),
+				'to' => $updateScheme['title'] ?? $table->getTitle(),
+			],
+			'emoji' => [
+				'from' => $table->getEmoji(),
+				'to' => $updateScheme['emoji'] ?? $table->getEmoji(),
+			],
+			'description' => [
+				'from' => $table->getDescription(),
+				'to' => $updateScheme['description'] ?? $table->getDescription(),
+			],
+			'columns' => [
+				'addColumns' => $structureService->addedColumns(),
+				'removeColumns' => $structureService->removedColumns(),
+				'modifyColumns' => $structureService->modifiedColumns(),
+			],
+			'views' => [
+				'addViews' => $structureService->addedViews(),
+				'removeViews' => $structureService->removedViews(),
+				'modifyViews' => $structureService->modifiedViews(),
+			],
+			'columnOrderChanges' => $structureService->columnOrderChanges(),
+			'sortChanges' => $structureService->sortChanges(),
+		];
+	}
+
+	/**
+	 * @param int $tableId
+	 * @param array $columns
+	 * @param array $views
+	 * @param array $columnOrder
+	 * @param array $sort
+	 * @param string|null $userId
+	 *
+	 * @return Table
+	 *
+	 * @throws InternalError
+	 * @throws NotFoundError
+	 * @throws PermissionError
+	 * @throws BadRequestError
+	 */
+	public function updateTableStructure(int $tableId, array $columns, array $views, array $columnOrder = [], array $sort = [], ?string $userId = null): Table
 	{
 		$userId = $this->permissionsService->preCheckUserId($userId);
 
@@ -637,21 +690,75 @@ class TableService extends SuperService {
 			throw new PermissionError('PermissionError: can not manage table with id ' . $tableId);
 		}
 
+		// Validate the structure of the columns and views arrays
+		if (!isset($columns['addColumns']) || !is_array($columns['addColumns'])
+			|| !isset($columns['removeColumns']) || !is_array($columns['removeColumns'])
+			|| !isset($columns['modifyColumns']) || !is_array($columns['modifyColumns'])) {
+			throw new BadRequestError('Invalid columns structure provided.');
+		}
+
+		if (!isset($views['addViews']) || !is_array($views['addViews'])
+			|| !isset($views['removeViews']) || !is_array($views['removeViews'])
+			|| !isset($views['modifyViews']) || !is_array($views['modifyViews'])) {
+			throw new BadRequestError('Invalid views structure provided.');
+		}
+
 		// Add new columns
-		foreach ($addColumns as $columnData) {
-			$this->columnService->create($userId, $tableId, null, ColumnDto::createFromArray($columnData));
+		foreach ($columns['addColumns'] as $columnData) {
+			$this->columnService->importColumn($table, $columnData);
 		}
 
 		// Remove columns
-		foreach ($removeColumns as $columnData) {
+		foreach ($columns['removeColumns'] as $columnData) {
 			$this->columnService->delete($columnData['id']);
 		}
 
 		// Update existing columns
-		foreach ($updateColumns as $columnData) {
+		foreach ($columns['modifyColumns'] as $columnData) {
 			$fromColumn = $columnData['from'];
 			$toColumn = $columnData['to'];
 			$this->columnService->update($fromColumn['id'], $userId, ColumnDto::createFromArray($toColumn));
 		}
+
+		$updatedColumns = $this->columnService->findAllByTable($tableId, $userId, $table);
+		$columnsMap = [];
+		foreach ($updatedColumns as $column) {
+			$columnsMap[$column->getUuid()] = $column;
+		}
+
+		// Update column order
+		$columnSettings = ColumnSettings::createFromInputArray($columnOrder, $columnsMap);
+		$table->setColumnOrder(\json_encode($columnSettings->jsonSerialize()));
+
+		// Update sort rules
+		$sortRuleSet = SortRuleSet::createFromInputArray($sort, $columnsMap);
+		$table->setSort(\json_encode($sortRuleSet->jsonSerialize()));
+
+		try {
+			$table = $this->mapper->update($table);
+		} catch (OcpDbException $e) {
+			$this->logger->error($e->getMessage(), ['exception' => $e]);
+			throw new InternalError(get_class($this) . ' - ' . __FUNCTION__ . ': ' . $e->getMessage());
+		}
+
+		// Add views
+		foreach ($views['addViews'] as $view) {
+			$this->viewService->importView($tableId, $view, $userId);
+		}
+
+		// Modify views
+		foreach ($views['modifyViews'] as $item) {
+			$fromView = $item['from'];
+			$toView = $item['to'];
+			$existingView = $this->viewService->find($fromView['id'], $userId);
+			$this->viewService->update($existingView->getId(), ViewUpdateInput::fromInputArray($toView, $columnsMap), $userId);
+		}
+
+		// Remove views
+		foreach ($views['removeViews'] as $viewData) {
+			$this->viewService->delete($viewData['id']);
+		}
+
+		return $table;
 	}
 }
